requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
httpx==0.27.0

# ─── Database PostgreSQL ───
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
alembic==1.13.2

# ─── Autenticazione TOTP ───
pyotp==2.9.0
qrcode[pil]==7.4.2

# ─── Google (opzionale, per sync Sheets) ───
gspread==6.0.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.122.0

# ─── Scheduler ───
apscheduler==3.10.4
